recommendation/requirements.txt
pandas
numpy
matplotlib
seaborn
Pillow
os-sys
tensorflow
streamlit
sklearn
